pass aws credentials via env

+++ deploy.sh
@@ -5,6 +5,16 @@ if [[ ! "${SQS_URL+set}" ]]; then
   exit 1
 fi
 
+if [[ ! "${AWS_ACCESS_KEY_ID+set}" ]]; then
+  echo "set env AWS_ACCESS_KEY_ID"
+  exit 1
+fi
+
+if [[ ! "${AWS_SECRET_ACCESS_KEY+set}" ]]; then
+  echo "set env AWS_SECRET_ACCESS_KEY"
+  exit 1
+fi
+
 docker-compose -f docker-compose.yml -f docker-compose.prod.yml ps
 docker-compose -f docker-compose.yml -f docker-compose.prod.yml build
 docker-compose -f docker-compose.yml -f docker-compose.prod.yml down
